requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.11
django-environ==0.11.2
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.1
factory-boy==3.3.0
Faker==25.0.1
inflection==0.5.1
iniconfig==2.0.0
packaging==24.0
pluggy==1.5.0
psycopg2==2.9.9
PyJWT==2.8.0
pytest==8.2.0
pytest-factoryboy==2.7.0
python-dateutil==2.9.0.post0
python-decouple==3.8
six==1.16.0
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.11.0
